Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/css/css-text/overflow-wrap/overflow-wrap-break-word-keep-all-001-expected.txt

0000000000
헬로우월드헬로우월드헬로우월드헬로우월드헬로우월드

FAIL 0000000000 assert_greater_than: expected a number greater than 20 but got 20
PASS 헬로우월드헬로우월드헬로우월드헬로우월드헬로우월드
